refactor(core): collapse the external message converter derivation chains
