+++ .changeset/external-message-converter-split.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/core": patch
+---
+
+refactor: collapse the external message converter's duplicate derivation chains into one pure module.
